requirements.txt
Flask==2.3.3
psycopg2==2.9.7
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator